Linear regression

+++ assignment1/mad.h
@@ -1,12 +1,11 @@
 #include <iostream>
 #include <math.h>
-#include "mean.h"
 
 using namespace std;
 
-double mad(int* &ar, int size, int meanNum){
-    double sum = 0;
+long double mad(int* &ar, int size, long double meanNum){
+    long double sum = 0;
     for (int i = 0; i < size; i++)
         sum = sum + abs(ar[i] - meanNum);
-    return sum / size;
+    return sum / (long double)size;
 }
